pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "elevenlabs>=2.30.0",
    "flask>=3.1.2",
    "httpx>=0.28.1",
    "moviepy>=2.2.1",
    "openai>=2.15.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "sendgrid>=6.12.5",
    "tweepy>=4.16.0",
]
